requirements.txt
Flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
reportlab==4.0.8
Pillow==10.1.0
uuid==1.30
Werkzeug==3.0.1
